+++ tool/culture_compiler/encode_dart.dart
@@ -37,7 +37,7 @@ import 'package:time_machine2/src/platforms/dart_native_io.dart';
 
 void registerCulturesAsset() {
   TimeMachineIO.registerAsset('cultures', 'cultures.bin',
-      ByteData.sublistView(Uint8List.fromList(_embeddedData.codeUnits)));
+      ByteData.sublistView(Uint16List.fromList(_embeddedData.codeUnits)));
 }
 
 const _embeddedData =
